+++ src/users/users.controller.ts
@@ -1,4 +1,5 @@
-import { Body, Controller, Delete, Get, Param, ParseUUIDPipe, Patch, Post } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, ParseUUIDPipe, Patch, Post, Request, UnauthorizedException, UseGuards } from '@nestjs/common';
+import { AuthGuard } from 'src/auth/auth.guard';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserEmailDto } from './dto/update-user-email.dto';
 import { UpdateUserPasswordDto } from './dto/update-user-password.dto';
@@ -26,31 +27,48 @@ export class UsersController {
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuard)
   async update(
+    @Request() request: Request,
     @Param('id', ParseUUIDPipe) id: string,
     @Body() updateUserDto: UpdateUserDto
   ): Promise<User> {
+    if (request['sub'] !== id) throw new UnauthorizedException();
+
     return this.usersService.update(id, updateUserDto);
   }
 
   @Patch(':id/email')
+  @UseGuards(AuthGuard)
   async updateEmail(
+    @Request() request: Request,
     @Param('id', ParseUUIDPipe) id: string,
     @Body() { email }: UpdateUserEmailDto
   ): Promise<User> {
+    if (request['sub'] !== id) throw new UnauthorizedException();
+
     return this.usersService.updateEmail(id, email);
   }
 
   @Patch(':id/password')
+  @UseGuards(AuthGuard)
   async updatePassword(
+    @Request() request: Request,
     @Param('id', ParseUUIDPipe) id: string,
     @Body() { password }: UpdateUserPasswordDto
   ): Promise<User> {
+    if (request['sub'] !== id) throw new UnauthorizedException();
+
     return this.usersService.updatePassword(id, password);
   }
 
   @Delete(':id')
-  async remove(@Param('id', ParseUUIDPipe) id: string): Promise<User> {
+  @UseGuards(AuthGuard)
+  async remove(
+    @Request() request: Request,
+    @Param('id', ParseUUIDPipe) id: string): Promise<User> {
+    if (request['sub'] !== id) throw new UnauthorizedException();
+
     return this.usersService.remove(id);
   }
 }
